fix: improve typings for navigation
