+++ src/constants/index.ts
@@ -1,4 +1,5 @@
 export * from './asyncStorageKeys'
+export * from './breakpoints'
 export * from './colorSchemes'
 export * from './env'
 export * from './environments'
@@ -6,4 +7,4 @@ export * from './NativeBaseConfig'
 export * from './regex'
 export * from './theme'
 export * from './time'
-export * from './breakpoints'
+export * from './navigation'
