+++ src/constants/navigation.ts
@@ -1,4 +1,4 @@
-import { theme } from '~constants'
+import { theme } from './theme'
 
 export const MOBILE_MIN_WIDTH = 320
 export const WEB_NAV_BAR_ICON_SIZE = 24
@@ -8,3 +8,5 @@ export const WEB_SCREEN_STYLES = {
   minWidth: MOBILE_MIN_WIDTH,
   width: '100%',
 } as const
+
+export const TAB_DEFAULT_ICON = 'alert-line'
