+++ src/navigation/HomeStack.tsx
@@ -1,28 +1,10 @@
 import { createStackNavigator } from '@react-navigation/stack'
-import { t } from 'i18next'
-import { DetailsScreen, HomeScreen } from '~screens'
+
+import { homeStackScreensData } from './config/screens'
 
 import { useMemo } from '~hooks'
 
 const { Navigator, Screen } = createStackNavigator<HomeStackParamList>()
-
-export const enum HomeStackScreens {
-  Home = 'Home',
-  Details = 'Details',
-}
-
-export const homeStackScreensData = [
-  {
-    name: HomeStackScreens.Home,
-    component: HomeScreen,
-    title: t('navigation.screen_titles.home'),
-  },
-  {
-    name: HomeStackScreens.Details,
-    component: DetailsScreen,
-    title: t('navigation.screen_titles.details'),
-  },
-]
 
 export const HomeStack = (): JSX.Element => {
   const screens = useMemo(
