+++ src/navigation/config/enums.ts
@@ -1,0 +1,19 @@
+// SCREENS
+export const enum ExamplesStackScreens {
+  Examples = 'Examples',
+  Typography = 'Typography',
+  Colors = 'Colors',
+  Components = 'Components',
+  DataFromBeScreen_EXAMPLE = 'DataFromBeScreen',
+}
+
+export const enum HomeStackScreens {
+  Home = 'Home',
+  Details = 'Details',
+}
+
+// BOTTOM TABS
+export enum BottomTabsScreens {
+  Home = 'HomeStack',
+  Examples = 'ExamplesStack',
+}
