--- src/navigation/constants.ts
@@ -1,1 +1,0 @@
-export const TAB_DEFAULT_ICON = 'alert-line'
