--- src/navigation/webNavigator/utils.ts
@@ -1,1 +1,0 @@
-
