+++ src/screens/TypographyScreen.tsx
@@ -1,6 +1,7 @@
 import { Center, Divider, Heading, Row, ScrollView, Switch, Text } from 'native-base'
 
-import { useColorMode, useTranslation } from '~hooks'
+import { useColorScheme } from '~contexts'
+import { useTranslation } from '~hooks'
 
 export const fontSizes = [
   'xs',
@@ -20,7 +21,8 @@ export const fontSizes = [
 export const headingSizes = ['xs', 'sm', 'md', 'lg', 'xl', '2xl', '3xl', '4xl']
 
 export const TypographyScreen = (): JSX.Element => {
-  const { toggleColorMode, colorMode } = useColorMode()
+  const { setColorSchemeSetting, colorScheme } = useColorScheme()
+
   const { t } = useTranslation()
   return (
     <ScrollView>
@@ -31,7 +33,12 @@ export const TypographyScreen = (): JSX.Element => {
             Investigate the issue about using `useCallback` on `onChange`
             https://github.com/adobe/react-spectrum/issues/2320
           */}
-          <Switch mx={4} my={8} value={colorMode === 'dark'} onChange={toggleColorMode} />
+          <Switch
+            mx={4}
+            my={8}
+            value={colorScheme === 'dark'}
+            onChange={() => setColorSchemeSetting(colorScheme === 'dark' ? 'light' : 'dark')}
+          />
           <Text>🌚</Text>
         </Row>
         <Text fontSize="4xl">{t('typography_screen.text_font_size')}</Text>
