+++ src/services/NotificationService.ts
@@ -25,12 +25,11 @@ export const registerForPushNotificationsAsync = async () => {
       finalStatus = status
     }
     if (finalStatus !== 'granted') {
-      alert('Failed to get push token for push notification!')
+      console.log('Failed to get push token for push notification!')
       throw new Error('Failed to get push token for push notification!')
     }
   } else {
-    alert('Must use physical device for Push Notifications')
-    throw new Error('Must use physical device for Push Notifications')
+    console.log('Must use physical device for Push Notifications')
   }
 }
