+++ src/utils/testUtils.tsx
@@ -6,6 +6,7 @@ import { I18nextProvider } from 'react-i18next'
 
 import i18n from '~i18n/i18nForTests'
 import { AuthProvider } from '~providers'
+import { ColorSchemeProvider } from '~providers/ColorSchemeProvider'
 
 type RenderOptions = Parameters<typeof render>[1]
 
@@ -18,7 +19,9 @@ const ProvidersWrapper: React.FC<PropsWithChildren> = ({ children }) => (
   <AuthProvider>
     <NativeBaseProvider initialWindowMetrics={nbInitialWindowMetrics}>
       <NavigationContainer>
-        <I18nextProvider i18n={i18n}>{children}</I18nextProvider>
+        <ColorSchemeProvider>
+          <I18nextProvider i18n={i18n}>{children}</I18nextProvider>
+        </ColorSchemeProvider>
       </NavigationContainer>
     </NativeBaseProvider>
   </AuthProvider>
